+++ src/storm/modelchecker/lexicographic/spotHelper/spotProduct.cpp
@@ -29,12 +29,12 @@ typedef std::vector<std::pair<unsigned, unsigned>> product_states;
 
 std::shared_ptr<storm::automata::DeterministicAutomaton> ltl2daSpotProduct(storm::logic::MultiObjectiveFormula const& formula,
                                                                            storm::logic::ExtractMaximalStateFormulasVisitor::ApToFormulaMap& extracted,
-                                                                           std::vector<uint>& acceptanceConditions) {
+                                                                           std::vector<uint64_t>& acceptanceConditions) {
 #ifdef STORM_HAVE_SPOT
     bool first = true;
     spot::twa_graph_ptr productAutomaton;
     spot::bdd_dict_ptr dict = spot::make_bdd_dict();
-    uint countAccept = 0;
+    uint64_t countAccept = 0;
     // iterate over all subformulae
     for (const std::shared_ptr<const storm::logic::Formula>& subFormula : formula.getSubformulas()) {
         // get the formula in the right format
